requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
requests
google-generativeai
imbalanced-learn